requirements.txt
glob3
numpy
opencv-python
Pillow
scikit-learn
scipy
tensorflow